Move file to another package
