+++ src/main/java/seedu/souschef/commons/events/ui/CrossRecipePanelSelectionChangedEvent.java
@@ -1,6 +1,6 @@
 package seedu.souschef.commons.events.ui;
 
-import seedu.souschef.model.shop.CrossRecipe;
+import seedu.souschef.model.recipe.CrossRecipe;
 
 public class CrossRecipePanelSelectionChangedEvent extends GenericPanelSelectionChangedEvent<CrossRecipe> {
     public CrossRecipePanelSelectionChangedEvent(CrossRecipe newSelection) {
