+++ src/main/java/seedu/souschef/logic/LogicManager.java
@@ -16,7 +16,7 @@ import seedu.souschef.model.healthplan.HealthPlan;
 import seedu.souschef.model.ingredient.Ingredient;
 import seedu.souschef.model.planner.Day;
 import seedu.souschef.model.recipe.Recipe;
-import seedu.souschef.model.shop.CrossRecipe;
+import seedu.souschef.model.recipe.CrossRecipe;
 import seedu.souschef.storage.Storage;
 import seedu.souschef.storage.StorageManager;
 import seedu.souschef.ui.Ui;
