+++ src/main/java/seedu/souschef/model/AppContent.java
@@ -13,7 +13,7 @@ import seedu.souschef.model.planner.Day;
 import seedu.souschef.model.planner.DayComparator;
 import seedu.souschef.model.planner.Meal;
 import seedu.souschef.model.recipe.Recipe;
-import seedu.souschef.model.shop.CrossRecipe;
+import seedu.souschef.model.recipe.CrossRecipe;
 import seedu.souschef.model.tag.Tag;
