+++ src/main/java/seedu/souschef/model/Model.java
@@ -7,7 +7,7 @@ import javafx.collections.ObservableList;
 import seedu.souschef.model.healthplan.HealthPlan;
 import seedu.souschef.model.ingredient.Ingredient;
 import seedu.souschef.model.recipe.Recipe;
-import seedu.souschef.model.shop.CrossRecipe;
+import seedu.souschef.model.recipe.CrossRecipe;
 
 
 /**
